Harden calendar sync and split text helpers into summary module
- Include description and url in event_state_hash so description-only
  changes refresh the calendar event
- BedrockSummarizer::from_env now validates AWS credentials instead of
  always returning Some
- next_date returns Result rather than silently passing invalid input
  through; event_times and event_body propagate the error
- Narrow Google OAuth scope to calendar.events
- Capture HTTP response body in error messages via a check_status helper
- Parallelize per-calendar sync with futures::try_join_all
- Move text-summary helpers (event_summary, location_phrase,
  first_sentence, etc.) and their tests into a new summary module
- Drop the unused SecretString wrapper and the src/mbta-client/ stub
- Cargo.toml: replace 1.* wildcards with 1; trim tokio features to
  macros + rt-multi-thread
