+++ Cargo.toml
@@ -19,12 +19,13 @@ aws-config = "1"
 aws-sdk-bedrockruntime = "1"
 chrono = "0.4"
 clap = { version = "4.6", features = ["env"] }
+futures = "0.3"
 gcp_auth = "0.12"
 jluszcz_rust_utils = { git = "https://github.com//jluszcz/rust-utils", features = ["query"] }
-lambda_runtime = "1.*"
+lambda_runtime = "1"
 log = "0.4"
 reqwest = { version = "0.13", features = ["gzip", "json"] }
 rustls = { version = "0.23", default-features = false }
 serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.*"
-tokio = { version = "1.*", features = ["full"] }
+serde_json = "1"
+tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
