--- src/mbta-client/.gitignore
@@ -1,9 +1,0 @@
-/target/
-**/*.rs.bk
-Cargo.lock
-/src/
-/docs/
-/.openapi-generator
-README.md
-git_push.sh
-Cargo.toml
