--- src/mbta-client/.openapi-generator-ignore
@@ -1,2 +1,0 @@
-.gitignore
-.travis.yml
